+++ fsl_mrs/tests/test_utils_quantify.py
@@ -60,6 +60,59 @@ def test_QuantificationInfo():
     assert qci.add_corr == 5.0
 
 
+def test_volumefraction_calc():
+    qci = quant.QuantificationInfo(0.010, 3, ['NAA'], 298)
+    qci.set_fractions({'GM': 0.45, 'WM': 0.40, 'CSF': 0.15})
+    assert qci.f_GM == 0.45
+    assert qci.f_WM == 0.40
+    assert qci.f_CSF == 0.15
+
+
+def test_molefraction_calc():
+    qci = quant.QuantificationInfo(0.010, 3, ['NAA'], 298)
+    qci.set_fractions({'GM': 0.45, 'WM': 0.40, 'CSF': 0.15})
+
+    # Densitites are 'GM': 0.78, 'WM': 0.65, 'CSF': 0.97
+    sum_frac = (0.45 * 0.78 + 0.40 * 0.65 + 0.15 * 0.97)
+    assert np.isclose(qci.f_GM_H2O, 0.45 * 0.78 / sum_frac)
+    assert np.isclose(qci.f_WM_H2O, 0.40 * 0.65 / sum_frac)
+    assert np.isclose(qci.f_CSF_H2O, 0.15 * 0.97 / sum_frac)
+
+
+def test_corrected_water_conc():
+    # No relaxation
+    qci = quant.QuantificationInfo(1E-10, 1E5, ['NAA'], 298)
+    qci.set_fractions({'GM': 1.00, 'WM': 0.0, 'CSF': 0.0})
+
+    print(qci.relax_corr_water_molal)
+    print(qci.relax_corr_water_molar)
+    # Molality should be close to pure water as density term cancels
+    assert np.isclose(qci.relax_corr_water_molal, 55510)
+    # Molarity should be scaled by density term as volume fixed
+    assert np.isclose(qci.relax_corr_water_molar, 55510 * 0.78)
+
+    qci.set_fractions({'GM': 0.50, 'WM': 0.5, 'CSF': 0.0})
+
+    print(qci.relax_corr_water_molal)
+    print(qci.relax_corr_water_molar)
+    # Molality should be close to pure water as density term cancels
+    assert np.isclose(qci.relax_corr_water_molal, 55510)
+    # Molarity should be scaled by density terms as volume fixed
+    assert np.isclose(qci.relax_corr_water_molar, 55510 * (0.78 + 0.65) / 2)
+
+    qci = quant.QuantificationInfo(1E-10, 1, ['NAA'], 298)
+    qci.set_fractions({'GM': 0.50, 'WM': 0.5, 'CSF': 0.0})
+
+    print(qci.relax_corr_water_molal)
+    print(qci.relax_corr_water_molar)
+    # Molality should scaled by relaxation terms in proportion of mole fraction.
+    mf_gm = 0.5 * 0.78 / (0.5 * 0.78 + 0.5 * 0.65)
+    mf_wm = 0.5 * 0.65 / (0.5 * 0.78 + 0.5 * 0.65)
+    assert np.isclose(qci.relax_corr_water_molal, 55510 * (qci.R_H2O_GM * mf_gm + qci.R_H2O_WM * mf_wm))
+    # Molarity should be scaled by density terms * relaxation terms
+    assert np.isclose(qci.relax_corr_water_molar, 55510 * (0.78 * qci.R_H2O_GM + 0.65 * qci.R_H2O_WM) / 2)
+
+
 def test_quantifyWater():
     basis = mrsio.read_basis(basisfile)
     data = mrsio.read_FID(metabfile)
@@ -107,3 +160,4 @@ def test_quantifyWater():
 
     assert np.allclose(res.getConc(scaling='internal'), 1.0)
     assert np.allclose(res.getConc(scaling='molarity'), 10.78, atol=1E-1)
+    assert np.allclose(res.getConc(scaling='molality'), 10.78 * 1 / (0.6 * 0.78 + 0.4 * 0.65), atol=1E-1)